backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

pydantic==2.4.2
pydantic-settings==2.0.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7

typing-extensions==4.8.0

supabase==2.3.0
postgrest==0.13.2
gotrue==2.4.2

openai==1.6.1
httpx==0.24.1